Rage cambiada para que se sume progresivamente.

+++ Goblin Slayer/Assets/Scripts/Health.cs
@@ -6,7 +6,7 @@ public class Health : MonoBehaviour
 {
 
     public int maxHealth;
-    public int currentHealth;
+    public float currentHealth;
     public bool Alive { get; private set; }
     private Animator animator;
 
@@ -20,7 +20,7 @@ public class Health : MonoBehaviour
     /// Applies damage and updates the gameObject's status (alive/dead)
     /// </summary>
     /// <param name="amount"> Amount of damage </param>
-    public void LoseHealth(int amount)
+    public void LoseHealth(float amount)
     {
         HitAnimator();
         currentHealth -= amount;
@@ -62,7 +62,7 @@ public class Health : MonoBehaviour
     /// Get current HP
     /// </summary>
     /// <returns></returns>
-    public int GetHP()
+    public float GetHP()
     {
         return currentHealth;
     }
